Create kafka consumer/producer per query
