+++ sqrl-planner/sqrl-planner-local/src/test/java/com/datasqrl/util/data/Sensors.java
@@ -11,10 +11,16 @@ public class Sensors extends UseCaseExample {
   public static final Sensors INSTANCE = new Sensors("");
 
   public static final Sensors INSTANCE_EPOCH = new Sensors("epoch");
-
+  public static final Sensors INSTANCE_MUTATION
+      = new Sensors("mutation", "metricsapi.graphqls");
 
   protected Sensors(String variant) {
+    this(variant, null);
+  }
+
+  protected Sensors(String variant, String graphql) {
     super(variant, Set.of("sensors","sensorreading","machinegroup"),
-        script("sensors-teaser","machine","minreadings"));
+        script("metrics","SecReading","minreadings"),
+        graphql);
   }
 }
