+++ sqrl-testing/sqrl-integration-tests/src/test/java/com/datasqrl/util/TestPackager.java
@@ -1,0 +1,41 @@
+package com.datasqrl.util;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.SneakyThrows;
+import org.apache.kafka.streams.integration.utils.EmbeddedKafkaCluster;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Map;
+
+public class TestPackager {
+
+  public static Path createPackageOverride(EmbeddedKafkaCluster kafka, PostgreSQLContainer testDatabase) {
+    Map overrideConfig = Map.of("engines",
+        Map.of("log", Map.of("connector",
+                Map.of("bootstrap.servers", "localhost:" + kafka.bootstrapServers().split(":")[1])),
+            "database", toDbMap(testDatabase)));
+    return writeJson(overrideConfig);
+  }
+
+  public static Map toDbMap(PostgreSQLContainer testDatabase) {
+    return Map.of("database", testDatabase.getDatabaseName(),
+        "dialect", "postgres",
+        "driver", testDatabase.getDriverClassName(),
+        "port", testDatabase.getMappedPort(5432),
+        "host", testDatabase.getHost(),
+        "user", testDatabase.getUsername(),
+        "password", testDatabase.getPassword(),
+        "url", testDatabase.getJdbcUrl());
+  }
+
+  @SneakyThrows
+  public static Path writeJson(Map overrideConfig) {
+    ObjectMapper mapper = SqrlObjectMapper.INSTANCE;
+
+    Path tmp = Files.createTempFile("pkjson", ".json");
+    mapper.writeValue(tmp.toFile(), overrideConfig);
+    return tmp;
+  }
+}
