+++ sqrl-testing/sqrl-integration-tests/src/test/resources/snapshots/com/datasqrl/graphql/SensorsTest/multipleSubscribersTest.txt
@@ -1,0 +1,5 @@
+{"HighTempAlert":{"sensorid":1,"temp":53.11666667}}
+{"HighTempAlert":{"sensorid":1,"temp":53.11666667}}
+{"HighTempAlert":{"sensorid":1,"temp":53.11666667}}
+{"HighTempAlert":{"sensorid":1,"temp":53.11666667}}
+{"HighTempAlert":{"sensorid":1,"temp":53.11666667}}
